Publish handle state accessor
In order to make decisions based on the consistent state view in a handle
it's necessary to have access to the _state as known by the handle_ and
not just the state at point in time of request.
This adds the necessary call and refactors some state transmution logic
out of exec.Container and into the base objects.